WIP: Fusion Parser move internally used Classes but not the Parser
i didn't move the parser because some more files will be changed and one could lose the overview.
